Introduce Foreman to replace DaemonSpawn
This works well, but may not be ideal for the future.
If every "event" has its own db file and own daemon process, the
Procfile may not be the best method of managing this.
The original approach would allow a new daemon to be spawned by the
application for every "event" it is tracking.